+++ docker/scripts/install_qctools.sh
@@ -16,7 +16,11 @@ fossil checkout e5723df2c0c85959  # 2.2.2
 
 ./waf configure --prefix=/usr/
 ./waf install
-ln -s /usr/bin/qctool_v2.2.2 /usr/bin/qctool
+mv /usr/bin/hptest_v2.2.2 /usr/bin/hptest
+mv /usr/bin/inthinnerator_v2.2.2 /usr/bin/inthinnerator
+mv /usr/bin/ldbird_v2.2.2 /usr/bin/ldbird
+mv /usr/bin/qctool_v2.2.2 /usr/bin/qctool
+mv /usr/bin/selfmap_v2.2.2 /usr/bin/selfmap
 
 # remove fossil used to build qctools
 apt-get purge \